requirements.txt
pymupdf==1.23.8
python-docx==1.1.2
python-arango==7.9.1
groq>=0.11.0
fastembed==0.3.6
python-dotenv==1.0.1
tenacity==8.3.0
numpy==1.26.4
tqdm==4.66.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
